// src/app/interfaces/Solucion/Solucion.ts
export interface Solucion {
    id?: string;
    codi: string;
    ran1_codi: string;
    descripcion: string;
    num_revision: string;
    pidm: string;
    ind_paquete: string;
    ind_script: string;
    ind_bd: string;
    ind_forma: string;
    OTROS: string;
    modulo: string;
    extension: string;
    url: string;
    fechaActividad: string;
    idUsuario: string;
  }
  